PGY006-EXAMEN-DJANGO-main/requirements.txt

# pip install -r requirements.txt

django == 4.0.4
pillow
django-admin-interface
django-crispy-forms
mysqlclient
pymysql
djangorestframework
requests